--- a/packages/opentelemetry-context-async-hooks/src/index.ts
+++ b/packages/opentelemetry-context-async-hooks/src/index.ts
@@ -15,3 +15,4 @@
  */
 
 export * from './AsyncHooksContextManager';
+export * from './AsyncLocalStorageContextManager';