## Which problem is this PR solving?

Performance impact of using the current nodejs default context manager.

## Short description of the changes

`async_hooks` especially with [promise executing tracking](https://nodejs.org/docs/latest-v14.x/api/async_hooks.html#async_hooks_promise_execution_tracking) enabled are known to have a significant performance impact https://github.com/nodejs/benchmarking/issues/181

Luckily there is a new API in node [AsyncLocalStorage](https://nodejs.org/docs/latest-v14.x/api/async_hooks.html#async_hooks_class_asynclocalstorage) (available since v13.10.0+ and v12.17.0+), which is supposed to be scoped more towards the ContextManager use-case and is more performant.`AsyncLocalStorage` recently received additional perf improvements https://github.com/nodejs/node/pull/32891 and has more planned https://github.com/nodejs/diagnostics/issues/376

As this is a fairly recent API, I'm sure this can't be accepted as is and should be seen more as a conversation starter.

Open question:

- Was `AsyncLocalStorage` already tested and discarded (all test pass, but I might miss something)
- Does it make sense to be added as a separate context manager?
- Would it make sense to feature detected `AsyncLocalStorage` and use this as default?