@dyladan 2020-06-17
Think @OlivierAlbertini should weigh in on this, but in general I am not against including a more performant context manager. As far as feature detection goes, I'm not sure how I feel about having the default behavior change based on which version of node you use, but I am open to being convinced otherwise.