@johanneswuerbach 2020-06-17
As a datapoint from our mildly complex express app on node `v14.4.0`.

```
# Without opentelemetry tracer

Requests      [total, rate, throughput]         569, 9.48, 9.47
Duration      [total, attack, wait]             1m0s, 1m0s, 91.537ms
Latencies     [min, mean, 50, 90, 95, 99, max]  79.78ms, 105.634ms, 99.757ms, 116.015ms, 133.831ms, 208.806ms, 1.102s

# With current ContextManager

Requests      [total, rate, throughput]         446, 7.43, 7.41
Duration      [total, attack, wait]             1m0s, 1m0s, 127.091ms
Latencies     [min, mean, 50, 90, 95, 99, max]  108.509ms, 134.938ms, 130.591ms, 147.889ms, 160.71ms, 246.351ms, 555.653ms

# With AsyncLocalStorage based ContextManager

Requests      [total, rate, throughput]         507, 8.44, 8.42
Duration      [total, attack, wait]             1m0s, 1m0s, 111.535ms
Latencies     [min, mean, 50, 90, 95, 99, max]  90.519ms, 118.767ms, 113.332ms, 137.837ms, 151.618ms, 216.908ms, 299.603ms
```