@vmarchaud 2020-06-18
@dyladan Think you meant to tag me since Olivier didn't work on this part ^^ 


> Was AsyncLocalStorage already tested and discarded (all test pass, but I might miss something)

We didn't yet test it in OT, someone already brought the suggestion [here](https://github.com/open-telemetry/opentelemetry-js/pull/926#issuecomment-619339836) however my main complain with it was simply that the API is still experimental (way more than `async_hooks` itself), at the time there were still [some](https://github.com/nodejs/node/pull/31950) [fixes](https://github.com/nodejs/node/pull/32085) [made](https://github.com/nodejs/node/pull/33189) to the API.

> Does it make sense to be added as a separate context manager?

This is mandatory for me since we support node 8, 10 and versions that doesn't have this API. Also it allows users to opt-in.

> Would it make sense to feature detected AsyncLocalStorage and use this as default?

It may in the future but i would be against for now for reasons cited above. I totally agree that this is the way forward (for both performance and correctness reasons) but i would prefer to wait to have real world usage feedback.

> async_hooks especially with promise executing tracking enabled are known to have a significant performance impact nodejs/benchmarking#181

Correct me if i'm wrong but AsyncLocalStorage does track promise too (see [this](https://github.com/nodejs/node/pull/33189)) with the PromiseHook API ? 

To sum everything i said, i'm totally fine having a `AsyncLocalStorageContextManager` implementation so user can try this new API out. However i would wait for @Qard to "finish" his works and real world usage to make it the default (when available ofc).

PS: On the benchmark i believe it would be better to make micro-benchark (see #1123) because other code (that get JIT'ed by V8) might influence the result.