@Qard 2020-06-18
AsyncLocalStorage is as stable as async_hooks at this point, if not more so due to the reduced surface area.

It uses async_hooks internally, so everything async_hooks can handle it can handle too. Through async_hooks it uses promise hooks, however I made some changes to skip the additional garbage collection tracking to emit destroy events for promises if there is no destroy hook present. AsyncLocalStorage  only uses the init hook and the executionAsyncResource function, so it has much simpler needs than the full async_hooks capabilities. It can function without much of the usual overhead of async_hooks.

Both async_hooks and AsyncLocalStorage have also had issues with thenables. I have a fix that has already landed in V8 and is just in the process of landing in Node.js now.

Beyond that, AsyncLocalStorage should be stable enough for regular use. I’m still working on some more changes to async_hooks, but those are largely internal-facing other than some performance improvements.