@vmarchaud 2020-06-23
@johanneswuerbach Looks great overall ! I will take a closer look tomorrow. 

I'm wondering whats the best approach to maintain both implementation, should we make two different package ? The switch between `asynchooks` and `asynclocalstorage` will be made by the `node-sdk` anyway so it doesn't really matter.