@johanneswuerbach 2020-06-23
@dyladan I didn't remove any tests, hiding whitespaces https://github.com/open-telemetry/opentelemetry-js/pull/1210/files?diff=split&w=1 should make this more easy to compare.

I tried to touch as least code as possible, should I maybe split this up into two commits:

a) adding the new base context manager
b) adding the AsyncLocalStorageContextManager

?