@vmarchaud 2020-06-23
@johanneswuerbach Wow i didnt know that you could remove the whitespace diff from github, thanks ! Should be the default ahah