@obecny 2020-06-25
> Sorry I’m not sure about the process, should i squash my commits now?

https://github.com/open-telemetry/opentelemetry-js/blob/master/CONTRIBUTING.md#general-merge-requirements

don't merge, please wait for maintainer