@dyladan 2020-06-25
>Sorry I’m not sure about the process, should i squash my commits now?

Commits are squashed during the merge process anyway so it isn't that important unless you want the message to say something specific.

>don't merge, please wait for maintainer

he doesn't even have permission to merge. Only approvers/maintainers do.